+++ CHANGELOG.md
@@ -20,6 +20,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Tests for "Sum of divisors".
 - Tests for "Numeral system".
 - Tests for "Flood fill example".
+- Tests for "5D chests".
 
 ## [1.15.0] - 2022-11-30
 ### Added